[Bronze V] Title: 이상한 기호, Time: 56 ms, Memory: 5752 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -1,2 +1,2 @@
-var input = Console.ReadLine().Split().Select(int.Parse).ToList();
+var input = Console.ReadLine().Split().Select(long.Parse).ToList();
 Console.WriteLine((input[0] + input[1])*(input[0]-input[1]));
